+++ pykeepass/entry.py
@@ -24,7 +24,8 @@ reserved_keys = [
     'IconID',
     'Times',
     'History',
-    'Notes'
+    'Notes',
+    'otp'
 ]
 
 # FIXME python2
@@ -32,7 +33,7 @@ reserved_keys = [
 class Entry(BaseElement):
 
     def __init__(self, title=None, username=None, password=None, url=None,
-                 notes=None, tags=None, expires=False, expiry_time=None,
+                 notes=None, otp=None, tags=None, expires=False, expiry_time=None,
                  icon=None, autotype_sequence=None, autotype_enabled=True,
                  element=None, kp=None):
 
@@ -55,6 +56,8 @@ class Entry(BaseElement):
                 self._element.append(E.String(E.Key('URL'), E.Value(url)))
             if notes:
                 self._element.append(E.String(E.Key('Notes'), E.Value(notes)))
+            if otp:
+                self._element.append(E.String(E.Key('otp'), E.Value(otp)))
             if tags:
                 self._element.append(
                     E.Tags(';'.join(tags) if type(tags) is list else tags)
@@ -63,9 +66,10 @@ class Entry(BaseElement):
                 E.AutoType(
                     E.Enabled(str(autotype_enabled)),
                     E.DataTransferObfuscation('0'),
-                    E.DefaultSequence(str(autotype_sequence))
+                    E.DefaultSequence(str(autotype_sequence) if autotype_sequence else '')
                 )
             )
+            # FIXME: include custom_properties in constructor
 
         else:
             assert type(element) in [_Element, Element, ObjectifiedElement], \
@@ -77,15 +81,33 @@ class Entry(BaseElement):
             self._element = element
 
     def _get_string_field(self, key):
+        """Get a string field from an entry
+
+        Args:
+            key (str): name of field
+
+        Returns:
+            (str or None): field value
+        """
+
         field = self._xpath('String/Key[text()="{}"]/../Value'.format(key), first=True)
         if field is not None:
             return field.text
 
-    def _set_string_field(self, key, value):
+    def _set_string_field(self, key, value, protected=True):
+        """Create or overwrite a string field in an Entry
+
+        Args:
+            key (str): name of field
+            value (str): value of field
+            protected (bool): mark whether the field should be protected in memory
+                in other tools.  This property is ignored in PyKeePass and all
+                fields are decrypted immediately upon opening the database.
+        """
         field = self._xpath('String/Key[text()="{}"]/..'.format(key), first=True)
         if field is not None:
             self._element.remove(field)
-        self._element.append(E.String(E.Key(key), E.Value(value)))
+        self._element.append(E.String(E.Key(key), E.Value(value, Protected=str(protected))))
 
     def _get_string_field_keys(self, exclude_reserved=False):
         results = [x.find('Key').text for x in self._element.findall('String')]
@@ -120,6 +142,7 @@ class Entry(BaseElement):
 
     @property
     def title(self):
+        """str: get or set entry title"""
         return self._get_string_field('Title')
 
     @title.setter
@@ -128,6 +151,7 @@ class Entry(BaseElement):
 
     @property
     def username(self):
+        """str: get or set entry username"""
         return self._get_string_field('UserName')
 
     @username.setter
@@ -136,6 +160,7 @@ class Entry(BaseElement):
 
     @property
     def password(self):
+        """str: get or set entry password"""
         return self._get_string_field('Password')
 
     @password.setter
@@ -144,6 +169,7 @@ class Entry(BaseElement):
 
     @property
     def url(self):
+        """str: get or set entry URL"""
         return self._get_string_field('URL')
 
     @url.setter
@@ -152,6 +178,7 @@ class Entry(BaseElement):
 
     @property
     def notes(self):
+        """str: get or set entry notes"""
         return self._get_string_field('Notes')
 
     @notes.setter
@@ -160,6 +187,7 @@ class Entry(BaseElement):
 
     @property
     def icon(self):
+        """str: get or set entry icon. See icons.py"""
         return self._get_subelement_text('IconID')
 
     @icon.setter
@@ -168,6 +196,7 @@ class Entry(BaseElement):
 
     @property
     def tags(self):
+        """str: get or set entry tags"""
         val = self._get_subelement_text('Tags')
         return val.split(';') if val else val
 
@@ -178,7 +207,17 @@ class Entry(BaseElement):
         return self._set_subelement_text('Tags', v)
 
     @property
+    def otp(self):
+        """str: get or set entry OTP text. (defacto standard)"""
+        return self._get_string_field('otp')
+
+    @otp.setter
+    def otp(self, value):
+        return self._set_string_field('otp', value)
+
+    @property
     def history(self):
+        """:obj:`list` of :obj:`HistoryEntry`: get entry history"""
         if self._element.find('History') is not None:
             return [HistoryEntry(element=x, kp=self._kp) for x in self._element.find('History').findall('Entry')]
         else:
@@ -190,6 +229,7 @@ class Entry(BaseElement):
 
     @property
     def autotype_enabled(self):
+        """bool: get or set autotype enabled state.  Determines whether `autotype_sequence` should be used"""
         enabled = self._element.find('AutoType/Enabled')
         if enabled.text is not None:
             return enabled.text == 'True'
@@ -204,8 +244,11 @@ class Entry(BaseElement):
 
     @property
     def autotype_sequence(self):
+        """str: get or set [autotype string](https://keepass.info/help/base/autotype.html)"""
         sequence = self._element.find('AutoType/DefaultSequence')
-        return sequence.text if sequence is not None else None
+        if sequence is None or sequence.text == '':
+            return None
+        return sequence.text
 
     @autotype_sequence.setter
     def autotype_sequence(self, value):
@@ -213,6 +256,7 @@ class Entry(BaseElement):
 
     @property
     def is_a_history_entry(self):
+        """bool: check if entry is History entry"""
         parent = self._element.getparent()
         if parent is not None:
             return parent.tag == 'History'
@@ -221,7 +265,7 @@ class Entry(BaseElement):
     @property
     def path(self):
         """Path to element as list.  List contains all parent group names
-        ending with entry title.  List may contain strings or NoneTypes."""
+        ending with entry title.  List contains strings or NoneTypes."""
 
         # The root group is an orphan
         if self.parentgroup is None:
@@ -234,9 +278,9 @@ class Entry(BaseElement):
             p = p.parentgroup
         return path
 
-    def set_custom_property(self, key, value):
+    def set_custom_property(self, key, value, protect=False):
         assert key not in reserved_keys, '{} is a reserved key'.format(key)
-        return self._set_string_field(key, value)
+        return self._set_string_field(key, value, protect)
 
     def get_custom_property(self, key):
         assert key not in reserved_keys, '{} is a reserved key'.format(key)
@@ -250,6 +294,25 @@ class Entry(BaseElement):
             raise AttributeError('Could not find property element')
         self._element.remove(prop)
 
+    def is_custom_property_protected(self, key):
+        """Whether a custom property is protected.
+
+        Return False if the entry does not have a custom property with the
+        specified key.
+
+        Args:
+            key (:obj:`str`): key of the custom property to check.
+
+        Returns:
+            bool: Whether the custom property is protected.
+
+        """
+        assert key not in reserved_keys, '{} is a reserved key'.format(key)
+        field = self._xpath('String/Key[text()="{}"]/../Value'.format(key), first=True)
+        if field is not None:
+            return field.attrib.get("Protected", "False") == "True"
+        return False
+
     @property
     def custom_properties(self):
         keys = self._get_string_field_keys(exclude_reserved=True)
@@ -259,7 +322,16 @@ class Entry(BaseElement):
         return props
 
     def ref(self, attribute):
-        """Create reference to an attribute of this element."""
+        """Create reference to an attribute of this element.
+
+        Args:
+            attribute (str): one of 'title', 'username', 'password', 'url', 'notes', or 'uuid'
+
+        Returns:
+            str: [field reference][fieldref] to this field of this entry
+
+        [fieldref]: https://keepass.info/help/base/fieldrefs.html
+        """
         attribute_to_field = {
             'title': 'T',
             'username': 'U',
@@ -272,7 +344,8 @@ class Entry(BaseElement):
 
     def save_history(self):
         '''
-        Save the entry in its history
+        Save the entry in its history.  History is not created unless this function is
+        explicitly called.
         '''
         archive = deepcopy(self._element)
         hist = archive.find('History')